+++ packages/core/src/models/index.ts
@@ -1,4 +1,8 @@
-export { ISSUE_ID_PATTERN, type IssueIdParts } from "./id";
+export {
+  ISSUE_ID_PATTERN,
+  parseIssueId,
+  type IssueIdParts,
+} from "./id";
 export {
   JIRA_RANK_MAPPED,
   computeReorderedRanks,